+++ src/common/mod.rs
@@ -1,3 +1,4 @@
 pub(crate) mod generate_nearest_n;
 pub(crate) mod generate_nearest_one;
 pub(crate) mod generate_within;
+pub(crate) mod generate_within_unsorted;
